[add] 로그아웃 함수 훅으로 분리
