+++ libs/hooks/src/lib/hooks.tsx
@@ -1,4 +1,7 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useCallback } from 'react';
+import { useQueryClient } from 'react-query';
+import { useRouter } from 'next/router';
+import { useLogoutMutation } from './mutation/useLogoutMutation';
 
 export function useSeconds() {
   const [value, setValue] = useState(0);
@@ -13,3 +16,20 @@ export function useSeconds() {
 }
 
 export default useSeconds;
+
+/**
+ * 로그아웃 함수
+ * - 프론트에서 글로벌에 저장하고 있는 유저정보 삭제
+ * - 로그아웃요청(토큰삭제)
+ */
+export function useLogout() {
+  const queryClient = useQueryClient();
+  const { mutateAsync } = useLogoutMutation();
+
+  const logout = useCallback(() => {
+    queryClient.removeQueries('Profile', { exact: true });
+    mutateAsync();
+  }, [mutateAsync, queryClient]);
+
+  return { logout };
+}
