+++ libs/hooks/src/lib/mutation/useLogoutMutation.tsx
@@ -1,6 +1,12 @@
+import { useRouter } from 'next/router';
 import { useMutation } from 'react-query';
 import axios from '../../axios';
 
 export const useLogoutMutation = () => {
-  return useMutation(() => axios.post('/auth/logout'));
+  const router = useRouter();
+  return useMutation(() => axios.post('/auth/logout'), {
+    onSuccess: () => {
+      router.push('/');
+    },
+  });
 };
